requirements.txt
# Core
streamlit==1.29.0
python-dotenv==1.0.0

# Document Processing
pdfplumber==0.10.3
pypdf==3.17.1
pytesseract==0.3.10
pdf2image==1.16.3

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.20

# Search
rank-bm25==0.2.2
scikit-learn==1.3.2

# LLM
llama-cpp-python==0.2.32

# Utils
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
Pillow==10.1.0

# Logging
loguru==0.7.2